+++ features/step_definitions/web_steps.rb
@@ -10,8 +10,9 @@ When 'I click on the header' do
   find('header a').click
 end
 
-When 'I select {string} from the state menu' do |state|
-  select state, from: 'State'
+When /^I (de)?select "(.+)" from the state menu$/ do |deselect, state|
+  method = deselect ? :unselect : :select
+  public_send method, state, from: 'State'
 end
 
 When /^I visit (.+)$/ do |page_name|
@@ -22,16 +23,20 @@ Then /^I should be on (.+)$/ do |page_name|
   expect(current_path).to be == path_to(page_name)
 end
 
-Then /^I should see a graph for (.+?) for the (\d+) day(?:s)? ending on (.+?)$/ do |state, days, end_date|
-  state = State.find state
+Then /^I should see a graph for (.+?) for the (\d+) day(?:s)? ending on (.+?)$/ do |states, days, end_date|
+  states = State.find states.strip.split(%r{\W+})
   end_date = Date.parse(end_date)
   start_date = end_date - (days.to_i.pred).days
   date_range = start_date..end_date
 
   page.find 'svg' do |svg|
-    expect(svg).to have_selector '.keyText', text: state.name
-    expect(svg).to have_xpath '//text[@class="dataPointPopup"][1]', text: /^#{start_date.strftime '%d %b %Y'},/
-    expect(svg).to have_xpath '//text[@class="dataPointPopup"][last()]', text:/^#{end_date.strftime '%d %b %Y'},/
+    states.each do |state|
+      expect(svg).to have_css '.keyText', text: state.name
+    end
+    expect(svg).to have_css '.dataPointLabel', count: states.count * days
+    [start_date, end_date].each do |date|
+      expect(svg).to have_css '.dataPointPopup', text: /^#{date.strftime '%d %b %Y'},/, count: states.count, visible: :all
+    end
   end
 end