+++ spec/models/chart_spec.rb
@@ -4,25 +4,24 @@ RSpec.describe Chart, type: :model do
   let(:legend) { Faker::Lorem.sentence }
 
   describe 'constructor' do
-    it 'takes an array of date-value pairs and a legend string' do
-      expect(described_class.new pairs: [[rand(100).days.ago, rand(100)]], legend: legend).to be_a_kind_of described_class
+    it 'takes a hash where the keys are legend strings and the values are arrays of date-value pairs' do
+      expect(described_class.new legend => [[rand(100).days.ago, rand(100)]]).to be_a_kind_of described_class
     end
   end
 
   describe 'instance methods' do
     let(:length) { rand 10..20 }
     let(:value_limit) { 100 }
-    let(:values) { Array.new(length) { rand value_limit } }
-    let(:dates) { Array.new(length) {|i| i.days.from_now.to_date}.shuffle }
-    let(:pairs) { dates.zip values }
-    let(:sorted_pairs) { pairs.sort }
-    let(:chart) { described_class.new(pairs: pairs, legend: legend) }
-
-    describe '#pairs' do
-      it 'returns the date-value pairs, sorted by date' do
-        expect(chart.pairs).to be == sorted_pairs
-      end
+    let(:series_count) { rand 2..5 }
+    let(:data) do
+      Array.new(series_count) do
+        dates = Array.new(length) {|i| i.days.from_now.to_date}.shuffle
+        values = Array.new(length) { rand value_limit }
+        [Faker::Lorem.sentence, dates.zip(values)]
+      end.to_h
     end
+
+    let(:chart) { described_class.new data }
 
     describe '#to_graph' do
       subject { chart.to_graph }
@@ -52,7 +51,7 @@ RSpec.describe Chart, type: :model do
         context 'divisions' do
           let(:divisions) { subject.scale_y_divisions }
 
-          before(:each) { values[rand values.length] = max_value if defined? max_value }
+          before(:each) { data[data.keys.sample][rand length][1] = max_value if defined? max_value }
 
           context 'max 100 or less' do
             it 'uses steps of 10' do
@@ -95,14 +94,27 @@ RSpec.describe Chart, type: :model do
       end
 
       context 'data' do
-        let(:data) { subject.instance_variable_get(:@data).first } # TODO: yes, this is terrible; let's see if we can do better
+        let(:svg_data) { subject.instance_variable_get(:@data) } # TODO: yes, this is terrible; let's see if we can do better
 
-        it 'puts the values into the data array, sorted by date and flattened' do
-          expect(data[:data]).to be == [sorted_pairs.map {|pair| DateTime.parse(pair.first.to_s).to_i }, sorted_pairs.map(&:last)]
+        it "puts each data series' values into the data array, sorted by date" do
+          expected = data.values.map do |pairs|
+            sorted_pairs = pairs.sort
+            [sorted_pairs.map {|(date, _)| DateTime.parse(date.to_s).to_i }, sorted_pairs.map {|(_, value)| value }]
+          end
+
+          expect(svg_data.pluck :data).to be == expected
         end
 
-        it 'uses the legend string as the name of the data series' do
-          expect(data[:title]).to be == legend
+        it 'uses each legend string as the name of the corresponding data series' do
+          expect(svg_data.pluck :title).to be == data.keys
+        end
+
+        context 'non-string keys' do
+          let(:data) { super().symbolize_keys }
+
+          it 'converts them to strings before using them as legends' do
+            expect(svg_data.pluck :title).to be == data.keys.map(&:to_s)
+          end
         end
       end
     end