+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased changes
 
+- Show multiple states on graph at once. [#37]
+
 # v1.3.1 / 2020-04-07
 
 - Run multiple requests through hydra. [#15]